requirements.txt
requests>=2.28.0
websocket-client>=1.6.0
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0
pydantic>=2.6.0
structlog>=24.1.0
python-dotenv>=1.0.0
pybit>=5.7.0
ccxt>=4.1.0
